gitea: refactor
